move steamcmd package into the main application
