+++ src/main/java/cz/forgottenempire/arma3servergui/common/util/SteamCmdWrapper.java
@@ -2,12 +2,12 @@ package cz.forgottenempire.arma3servergui.common.util;
 
 import cz.forgottenempire.arma3servergui.workshop.entities.DownloadStatus;
 import cz.forgottenempire.arma3servergui.workshop.entities.ErrorStatus;
-import cz.forgottenempire.steamcmd.SteamCmdExecutor;
-import cz.forgottenempire.steamcmd.SteamCmdParameters;
-import cz.forgottenempire.steamcmd.exceptions.IOOperationException;
-import cz.forgottenempire.steamcmd.exceptions.LoginException;
-import cz.forgottenempire.steamcmd.exceptions.NoMatchException;
-import cz.forgottenempire.steamcmd.exceptions.NoSubscriptionException;
+import cz.forgottenempire.arma3servergui.steamcmd.SteamCmdExecutor;
+import cz.forgottenempire.arma3servergui.steamcmd.SteamCmdParameters;
+import cz.forgottenempire.arma3servergui.steamcmd.exceptions.IOOperationException;
+import cz.forgottenempire.arma3servergui.steamcmd.exceptions.LoginException;
+import cz.forgottenempire.arma3servergui.steamcmd.exceptions.NoMatchException;
+import cz.forgottenempire.arma3servergui.steamcmd.exceptions.NoSubscriptionException;
 import java.io.File;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
